Basic framework
Added index.html which contains the basic framework for the website alongside the navigation bar.
Added script.js which contains the toggle for the mobile navigation bar section.
Added style.css which contains the styling as well as the media screen limits to switch from desktop to mobile navigation bars.