Update About Me Section
Modified index.html with addition of About Me section found by #about.
Modified style.css with formation of About Me section.
Added the necessary files for assets that came along with the section, along with a new logo to replace the plain word logo previously in the website.
Deleted unnecessary file in assets.